Work on eval. Create barcharts for each maze.
Fix bug where self.stats was returning a garbage episode (0 steps, from
latest reset), and flushing it out, so that the next eval episode on the
corresponding world would not be logged to self.stats.

Potentially fixed the number of eval episodes during train. Though this
may still be cluttered with some garbage episode stats.